Remove mapper-upload command and organize mapper cmds dir
