--- src/cmd/mapper/config.go
@@ -1,8 +1,0 @@
-package mapper
-
-const (
-	NamespacesKey       = "namespaces"
-	NamespacesShorthand = "n"
-	EnvIDKey            = "env-id"
-	SourceKey           = "source"
-)
