+++ src/cmd/mapper/mapper.go
@@ -1,6 +1,9 @@
 package mapper
 
 import (
+	"github.com/otterize/otterize-cli/src/cmd/mapper/export"
+	"github.com/otterize/otterize-cli/src/cmd/mapper/list"
+	"github.com/otterize/otterize-cli/src/cmd/mapper/reset"
 	"github.com/otterize/otterize-cli/src/pkg/mapperclient"
 	"github.com/spf13/cobra"
 )
@@ -16,8 +19,7 @@ func init() {
 	MapperCmd.PersistentFlags().String(mapperclient.MapperNamespaceKey, mapperclient.MapperNamespaceDefault, "the namespace of the mapper service")
 	MapperCmd.PersistentFlags().Int(mapperclient.MapperServicePortKey, mapperclient.MapperServicePortDefault, "the port of the mapper service")
 
-	MapperCmd.AddCommand(ExportCmd)
-	MapperCmd.AddCommand(ListCmd)
-	MapperCmd.AddCommand(ResetCmd)
-	MapperCmd.AddCommand(UploadCmd)
+	MapperCmd.AddCommand(export.ExportCmd)
+	MapperCmd.AddCommand(list.ListCmd)
+	MapperCmd.AddCommand(reset.ResetCmd)
 }
