+++ src/cmd/mapper/reset/mapper-reset.go
@@ -1,4 +1,4 @@
-package mapper
+package reset
 
 import (
 	"context"
